src/python/requirements.txt
protobuf>=3.5.0
pyzmq>=16.0.3
docutils>=0.12
nose=1.3.7
